requirements.txt
Flask==2.0.3
gunicorn==20.1.0
python-binance
google-cloud-secret-manager
